api/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
passlib==1.7.4
python-jose==3.3.0
mysqlclient==2.2.0
requests==2.31.0
beautifulsoup4==4.12.2
sqlalchemy==2.0.25
pydantic-settings==2.1.0
python-multipart==0.0.20
mysql-connector-python==9.3.0
argon2-cffi==21.3.0